feat: refactor cli(#149)
