--- cmd/cli/flydb.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"github.com/ByteStorage/FlyDB/cmd"
-	"github.com/desertbit/grumble"
-)
-
-func main() {
-	grumble.Main(cmd.App)
-}
