--- cmd/client/client.go
@@ -1,6 +1,0 @@
-package client
-
-// startDbClient starts a db client.
-func startDbClient() {
-
-}
