+++ cmd/client/root.go
@@ -1,28 +1,8 @@
-package cmd
+package client
 
-import (
-	"github.com/desertbit/grumble"
-)
+import "github.com/desertbit/grumble"
 
 func register(app *grumble.App) {
-	app.AddCommand(&grumble.Command{
-		Name: "start",
-		Help: "start server",
-		Run:  startServer,
-	})
-
-	app.AddCommand(&grumble.Command{
-		Name: "stop",
-		Help: "stop server",
-		Run:  stopServer,
-	})
-
-	app.AddCommand(&grumble.Command{
-		Name: "clean",
-		Help: "clean server",
-		Run:  cleanServer,
-	})
-
 	app.AddCommand(&grumble.Command{
 		Name: "put",
 		Help: "put data",
@@ -56,5 +36,4 @@ func register(app *grumble.App) {
 		Help: "list keys",
 		Run:  getKeys,
 	})
-
 }
