+++ cmd/server/server.go
@@ -1,6 +1,25 @@
 package server
 
-// startDbServer starts a db server.
-func startDbServer() {
+import (
+	"fmt"
+	"github.com/ByteStorage/FlyDB/config"
+	"github.com/ByteStorage/FlyDB/flydb"
+)
 
+func StartServer() {
+	options := config.DefaultOptions
+	_, err := flydb.NewFlyDB(options)
+	if err != nil {
+		fmt.Println("flydb start error: ", err)
+		return
+	}
+	fmt.Println("flydb start success")
+}
+
+func StopServer() {
+	panic("implement me")
+}
+
+func CleanServer() {
+	panic("implement me")
 }
